deployment/gradio-app/requirements.txt
gradio
fastai
torch
torchvision
opencv-python-headless
Pillow
numpy
matplotlib
scikit-image
